OldVersions/InstgramTheftyPosterV2/requirements.txt
instagrapi==1.17.0
openai==1.35.13
pyyaml
requests==2.28.1
rich==10.15.2
pillow
browser-cookie3
moviepy==1.0.3
cryptography==41.0.1
